requirements.txt
GitPython
python-gitlab
rich
typer
pymongo